# swagger.yaml
...
swagger: 2.0
info:
  description: Used in MyRA
  version: 1.0.0
  title: MyRA API
  contact:
    email: bcdevexchange@gov.bc.ca
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /AidanLaing/myra_api/1.0.0
...
schemes:
  - https
paths:
  /v1/agreement:
    get:
      tags:
        - Agreement
      summary: Returns array of all agreements
      description: Returns all agreements with nested data.
      operationId: readAgreements
      produces:
        - application/json
      parameters: []
      responses:
        200:
          description: Sucessfully returned agreements
          schema:
            type: array
            items:
              $ref: '#/definitions/Agreement'
        401:
          description: Not authorized.
      security:
        - bearerAuth: []
    post:
      tags:
        - Agreement
      summary: Creates an agreement
      description: Creates an agreement and returns the created object.
      operationId: createAgreement
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Agreement object needed to create agreement.
          required: true
          schema:
            $ref: '#/definitions/Agreement'
      responses:
        200:
          description: Agreement successfully created.
          schema:
            $ref: '#/definitions/Agreement'
        400:
          description: Bad or no body sent.
        401:
          description: Not authorized.
      security:
        - bearerAuth: []
  /v1/agreement/{agreementId}:
    get:
      tags:
        - Agreement
      summary: Returns agreement matching id
      description: Returns the agreement matching the id with nested data.
      operationId: readAgreement
      produces:
        - application/json
      parameters:
        - name: agreementId
          in: path
          description: Id of agreement to return
          required: true
          type: string
      responses:
        200:
          description: Successfully returned agreement matching id.
          schema:
            $ref: '#/definitions/Agreement'
        400:
          description: Invalid id supplied
        401:
          description: Not authorized.
        404:
          description: Agreement not found
      security:
        - bearerAuth: []
    put:
      tags:
        - Agreement
      summary: Update an agreement
      description: Updates an existing agreement and returns updated object.
      operationId: updateAgreement
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: agreementId
          in: path
          description: Id of agreement to update.
          required: true
          type: string
        - in: body
          name: body
          description: Agreement object to update existing agreement.
          required: true
          schema:
            $ref: '#/definitions/Agreement'
      responses:
        200:
          description: Successfully updated agreement.
          schema:
            $ref: '#/definitions/Agreement'
        400:
          description: Invalid id supplied.
        401:
          description: Not authorized.
        404:
          description: Agreement not found.
      security:
        - bearerAuth: []
...
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
definitions:
  Agreement:
    type: object
    required:
      - agreementEndDate
      - agreementStartDate
      - id
    properties:
      id:
        type: string
        example: RAN072542
      agreementStartDate:
        type: string
        example: 2018-03-15T16:52:36.961Z
      agreementEndDate:
        type: string
        example: 2043-03-09T16:52:36.961Z
      zoneId:
        type: integer
        example: 29
      typeId:
        type: integer
        example: 1
      exemptionStatusId:
        type: integer
        example: 1
      createdAt:
        type: string
        example: 2018-03-15T16:52:36.983Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:36.983Z
      zone:
        $ref: '#/definitions/Zone'
      livestockIdentifiers:
        type: array
        items:
          $ref: '#/definitions/LivestockIdentifier'
      plans:
        type: array
        items:
          $ref: '#/definitions/Plan'
      primaryAgreementHolder:
        $ref: '#/definitions/Client'
      usage:
        type: array
        items:
          $ref: '#/definitions/Usage'
    example:
      exemptionStatusId: 1
      usage:
        - temporaryIncrease: 0
          createdAt: 2018-03-15T16:52:37.658Z
          authorizedAum: 125
          totalAnnualUse: 125
          year: 2013
          agreementId: RAN077054
          totalNonUse: 0
          id: 7756
          updatedAt: 2018-03-15T16:52:37.658Z
        - temporaryIncrease: 0
          createdAt: 2018-03-15T16:52:37.658Z
          authorizedAum: 125
          totalAnnualUse: 125
          year: 2013
          agreementId: RAN077054
          totalNonUse: 0
          id: 7756
          updatedAt: 2018-03-15T16:52:37.658Z
      livestockIdentifiers:
        - createdAt: 2018-03-15T16:52:37.658Z
          livestockIdentifierType:
            createdAt: 2018-03-15T16:52:37.658Z
            description: Tag
            active: true
            id: 127412
            updatedAt: 2018-03-15T16:52:37.658Z
          livestockIdentifierLocation:
            createdAt: 2018-03-15T16:52:37.658Z
            description: Left Ear
            active: true
            id: 127412
            updatedAt: 2018-03-15T16:52:37.658Z
          agreementId: RAN1902841
          description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
          accepted: true
          id: 127412
          imageRef: TODO
          updatedAt: 2018-03-15T16:52:37.658Z
        - createdAt: 2018-03-15T16:52:37.658Z
          livestockIdentifierType:
            createdAt: 2018-03-15T16:52:37.658Z
            description: Tag
            active: true
            id: 127412
            updatedAt: 2018-03-15T16:52:37.658Z
          livestockIdentifierLocation:
            createdAt: 2018-03-15T16:52:37.658Z
            description: Left Ear
            active: true
            id: 127412
            updatedAt: 2018-03-15T16:52:37.658Z
          agreementId: RAN1902841
          description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
          accepted: true
          id: 127412
          imageRef: TODO
          updatedAt: 2018-03-15T16:52:37.658Z
      createdAt: 2018-03-15T16:52:36.983Z
      zone:
        createdAt: 2018-03-15T16:52:37.658Z
        code: DOS3
        districtId: 124124
        contactEmail: robjimmy@gov.bc.ca
        contactName: Rob Jimmy
        district:
          createdAt: 2018-03-15T16:52:37.658Z
          code: BEAV
          description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
          id: 15
          updatedAt: 2018-03-15T16:52:37.658Z
        description: Lakes TSA South of Francois Lake
        contactPhoneNumber: 1234567890
        id: 58
        updatedAt: 2018-03-15T16:52:37.658Z
      plans:
        - extension:
            createdAt: 2018-03-15T16:52:37.658Z
            notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
            endDate: 2018-03-15T16:52:37.658Z
            requestDate: 2018-03-15T16:52:37.658Z
            id: 1294812
            updatedAt: 2018-03-15T16:52:37.658Z
          notes: Example string of notes.
          pastures:
            - allowableAum: 400
              pdlPercent: 42.56
              createdAt: 2018-03-15T16:52:37.658Z
              notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
              graceDays: 12
              name: Sauer Mountain
              planId: 1
              id: 1
              updatedAt: 2018-03-15T16:52:37.658Z
            - allowableAum: 400
              pdlPercent: 42.56
              createdAt: 2018-03-15T16:52:37.658Z
              notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
              graceDays: 12
              name: Sauer Mountain
              planId: 1
              id: 1
              updatedAt: 2018-03-15T16:52:37.658Z
          createdAt: 2043-03-09T16:52:36.961Z
          grazingSchedules:
            - createdAt: 2018-03-15T16:52:37.658Z
              year: 2018
              narrative: TODO example
              id: 1092741
              updatedAt: 2018-03-15T16:52:37.658Z
              grazingScheduleEntries:
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
            - createdAt: 2018-03-15T16:52:37.658Z
              year: 2018
              narrative: TODO example
              id: 1092741
              updatedAt: 2018-03-15T16:52:37.658Z
              grazingScheduleEntries:
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
          statusId: 123456
          rangeName: Hudson, Barrows and Effertz
          agreementId: RAN077054
          id: 1
          planStartDate: 2043-03-09T16:52:36.961Z
          planEndDate: 2043-03-09T16:52:36.961Z
          extensionId: 123456
          updatedAt: 2043-03-09T16:52:36.961Z
          status:
            code: D
            name: Draft
            active: true
            id: 123982
        - extension:
            createdAt: 2018-03-15T16:52:37.658Z
            notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
            endDate: 2018-03-15T16:52:37.658Z
            requestDate: 2018-03-15T16:52:37.658Z
            id: 1294812
            updatedAt: 2018-03-15T16:52:37.658Z
          notes: Example string of notes.
          pastures:
            - allowableAum: 400
              pdlPercent: 42.56
              createdAt: 2018-03-15T16:52:37.658Z
              notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
              graceDays: 12
              name: Sauer Mountain
              planId: 1
              id: 1
              updatedAt: 2018-03-15T16:52:37.658Z
            - allowableAum: 400
              pdlPercent: 42.56
              createdAt: 2018-03-15T16:52:37.658Z
              notes: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
              graceDays: 12
              name: Sauer Mountain
              planId: 1
              id: 1
              updatedAt: 2018-03-15T16:52:37.658Z
          createdAt: 2043-03-09T16:52:36.961Z
          grazingSchedules:
            - createdAt: 2018-03-15T16:52:37.658Z
              year: 2018
              narrative: TODO example
              id: 1092741
              updatedAt: 2018-03-15T16:52:37.658Z
              grazingScheduleEntries:
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
            - createdAt: 2018-03-15T16:52:37.658Z
              year: 2018
              narrative: TODO example
              id: 1092741
              updatedAt: 2018-03-15T16:52:37.658Z
              grazingScheduleEntries:
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
                - createdAt: 2018-03-15T16:52:37.658Z
                  dateIn: 2018-03-15T16:52:37.658Z
                  livestockCount: 27
                  endDate: 2018-03-15T16:52:37.658Z
                  graceDays: 12
                  dateOut: 2018-03-15T16:52:37.658Z
                  livestockType:
                    auFactor: 0.1
                    createdAt: 2018-03-15T16:52:37.658Z
                    name: Alpaca
                    active: true
                    id: 127412
                    updatedAt: 2018-03-15T16:52:37.658Z
                  id: 129482
                  startDate: 2018-03-15T16:52:37.658Z
                  updatedAt: 2018-03-15T16:52:37.658Z
          statusId: 123456
          rangeName: Hudson, Barrows and Effertz
          agreementId: RAN077054
          id: 1
          planStartDate: 2043-03-09T16:52:36.961Z
          planEndDate: 2043-03-09T16:52:36.961Z
          extensionId: 123456
          updatedAt: 2043-03-09T16:52:36.961Z
          status:
            code: D
            name: Draft
            active: true
            id: 123982
      agreementStartDate: 2018-03-15T16:52:36.961Z
      agreementEndDate: 2043-03-09T16:52:36.961Z
      zoneId: 29
      typeId: 1
      id: RAN072542
      primaryAgreementHolder:
        createdAt: 2018-03-15T16:52:37.658Z
        clientType:
          createdAt: 2018-03-15T16:52:37.658Z
          code: A
          description: Licensee
          active: true
          id: 127412
          updatedAt: 2018-03-15T16:52:37.658Z
        clientEndDate: 2018-03-15T16:52:37.658Z
        name: Eriberto Fisher V
        location: XX
        id: 127412
        clientStartDate: 2018-03-15T16:52:37.658Z
        updatedAt: 2018-03-15T16:52:37.658Z
      updatedAt: 2018-03-15T16:52:36.983Z
...
  Zone:
    type: object
    required:
      - code
      - description
      - districtId
    properties:
      id:
        type: integer
        example: 58
      code:
        type: string
        example: DOS3
      description:
        type: string
        example: Lakes TSA South of Francois Lake
      contactName:
        type: string
        example: Rob Jimmy
      contactPhoneNumber:
        type: string
        example: 1234567890
      contactEmail:
        type: string
        example: robjimmy@gov.bc.ca
      districtId:
        type: integer
        example: 124124
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      district:
        $ref: '#/definitions/District'
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      code: DOS3
      districtId: 124124
      contactEmail: robjimmy@gov.bc.ca
      contactName: Rob Jimmy
      district:
        createdAt: 2018-03-15T16:52:37.658Z
        code: BEAV
        description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
        id: 15
        updatedAt: 2018-03-15T16:52:37.658Z
      description: Lakes TSA South of Francois Lake
      contactPhoneNumber: 1234567890
      id: 58
      updatedAt: 2018-03-15T16:52:37.658Z
  District:
    type: object
    properties:
      id:
        type: integer
        example: 15
      code:
        type: string
        example: BEAV
      description:
        type: string
        example: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
      createdAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
      updatedAt:
        type: string
        example: 2018-03-15T16:52:37.658Z
    example:
      createdAt: 2018-03-15T16:52:37.658Z
      code: BEAV
      description: Lorem ipsum dolor sit amet, consectetur adipiscing elit.
      id: 15
      updatedAt: 2018-03-15T16:52:37.658Z
...